convert ingest.py into a Django command
to test command:

```vagrant@ubuntu-bionic:/vagrant-test/maio$ python3 manage.py ingest```
